requirements.txt
strands-agents
anthropic
boto3
argparse
